crypto/hmac: speed up
Speed up hmac.Reset by avoiding recomputing the first block of the inner
and outer hash functions. For small messages, this saves two out of four
block computations and is a significant win. For longer messages, it's
not a big win but still a win.

As an add-on benefit, speeds up x/crypto/pbkdf2.

DO NOT SUBMIT

(Admittedly, this is patch is pretty horrible -- i'm open to better
suggestions. My first attempt used a Clone() hash.Hash method instead of
the Set(hash.Hash) method, which was slightly nicer but but added 2
allocations per Reset, which is not desirable.)

(This patch only affects HMAC-SHA256 for now. I can add the other hashes
in the stdlib if the overall approach is accepted.)